Strip BOM in CPDF_NameTree
PDFium doesn't strip BOMs during parsing, but we should strip BOMs when
retrieving parsed strings in CPDF_NameTree to ensure consistency and
appropriate function behavior. See the bug for more info.

As outlined in Bug=pdfium:593, the solution is to call GetUnicodeText()
.UTF8Encode() instead of GetString(). I added a GetUnicodeTextAt()
function in CPDF_Array, which is symmetrical to GetUnicodeTextFor() in
CPDF_Dictionary.

I also added a unit test for nametree, which would fail prior to this
change. Nametrees with non-unicode names are already tested by embedder
tests.

Bug=pdfium:820

Change-Id: Id69d7343632f83d1f5180348c0eea290f478183f