Add FORM_ExtendSelectionAndDelete() and embedder tests.
This method deletes the current text selection in a form text field or
user-editable form combobox text field plus the specified number of characters
before and after the selection. If there is no selection, the specified
number of characters before and after the current caret position are deleted.

This method will be used to cut text, and will be part of the PDF
plugin's implementation of WebPlugin::ExtendSelectionAndDelete().

BUG=chromium:59266

Change-Id: Ia3badca0129375ac4f6fb775dbb16c46f49b41c0