Add FPDFAnnot_GetInteractiveFormAnnotAtPoint() and embedder tests.
Given a point on a page, this method returns the annotation whose
rectangle contains that point. If there is no such annotation, nullptr
is returned. FPDFPage_CloseAnnot() must be called after the annotation
this method returns is no longer needed.

The tests for this method use the returned annotation to check its
interactive form annotation flags via
FPDFAnnot_GetInteractiveFormFlags().

BUG=chromium:59266

Change-Id: I8728792bea5367c172e08fdb6bae83eafb70eb68