DNS: Big rewrite of how FormCalc lexing logic works
This CL is not intended to be reviewed/submitted as is. I wanted to
upload it so that a) it isn't singly hosted on my workstation, and b)
have it available for referencing.

I am planning on breaking this is up into smaller/more manageable
parts, and eventually abandoning this once it is all landed.

I am not rebasing this as other CLs land, so some reviewer comments
like not using NUL termination are not reflected in it.

Change-Id: I4d5ae7b8881d15f7f4c501cf812b0c8853e92803