Change input to FMLexer to CFXA_WideString
This CL changes the input into the lexer to being a CFXA_WideString
instead of a CFXA_WideStringC. This in turn means that the lexer code
can depend on the fact that the string is null-terminated, instead of
having to perform error prone end pointer checks.

This is the first in a series of patches that are meant to cleanup the
existing lexer code and change how the error flow handling works. See
https://pdfium-review.googlesource.com/c/8570/ for context on where
this work is going.

BUG=pdfium:814

Change-Id: I533882df4031c6a683f527da43f9c1b14c5f1743