Update patch set 1
Patch Set 1: Code-Review-1

> Patch Set 1: Code-Review-1
> 
> What about embedded NULs?  
> 
> Generally, anything with an explicit length is much preferred over the NUL termination from a security/quality point of view. NUL termination must have seemed like a quaint idea in 1970, but has been a bane of our existence ever since.

The  FormCalc grammar doesn't include NUL as a valid character, so embedded NULs shouldn't be an issue. The current implementation of the lexer does include them in the valid set, but in general the only action taken on them is to return EOF, though I will admit that is probably not consistently done.

I don't really feel strongly about this change. It was introduced in my uber-patch as a consequence of there being a lot of code where the lexer extracts a character and works with that, instead of just dereferencing the cursor itself. I have removed that pattern, so on reflection there is probably limited utility in this change.

Patch-set: 1
Reviewer: Ryan Harrison <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Code-Review=-1
Label: Commit-Queue=0