Remove extracting characters from input
Through out the lexer there is use of a pattern where a character is
extracted from the location currently being pointed at in the input
and tests are done on this value. This is uneeded, since normally the
pointer isn't being modified while the character is stored, so the
pointer can be dereferenced directly.

This CL also cleans ups the implementation of string extraction from
the input to be easier to read, since it did depend on the character
being extracted to work. The IsFooCharacter methods are also
normalized to all take in a single character instead of some taking
pointers.

BUG=pdfium:814

Change-Id: I8eda01e0af59ff1f6c6e97d6fb504856c7a08a24