Clean up data passing in FormCalc Lexer
This CL removes the pattern used in the lexer of passing the lexing
member variables around as args to methods. Instead it uses the fact
that they are member variables in the methods.

This CL also includes renaming of variable and function names to
remove unneeded details or make them more precise.

BUG=pdfium:814

Change-Id: Id4c592338db9ff462835314252d39ab3b4b2b2ab