Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (2 comments)
> 
> The driver destructor is currently called from FPDF_RenderPage_Close on the next line; the destructor in turn calls Flush(). Not sure I grok how it works, but pPage->SetRenderContext(nullptr) replaces m_pRenderContext which triggers destructors all the way up to CFX_SkiaDeviceDriver.
> 
> If this isn't enough, let me know and I'll try my hand at re-plumbing it.

Is there any harm in calling Flush() twice? It would be nice to see it in the _Close call just for symmetry to the other calls in the progressive file so we can see what's happening without depending on the other chain.

I think we can leave SetRenderContext doing whatever it's doing and just add the Flush call?

Patch-set: 1
Reviewer: dsinclair <5965@fca2d64a-0420-3532-b200-12848c95a47b>