Add a build target and a proper export header for shared library builds.
This CL adds the pdf_is_shared_lib build target to pdfium. This is done
by adding a "pdf_component" template to the build files which provides
the same mechanisms as Chromiums component template, which is currently
unusable because Pdfium needs "is_component_build" to be set to false to
build correctly.

The export header stub in fpdfview.h is expanded to match Chromiums
export mechanisms and patterns (fixes pdfium:825).

Additionally, this CL provides a "pdf_source_set" template which
dynamically enables the use of source_set when building a complete
static library or shared library to reduce build time.

Bug: 825,826
Change-Id: Icedc538ec535e11d1e53c4d5fabc8c064b275752