Add ClickOnFormFieldAtPoint() helper to FPDFFormFillEmbeddertest.
This helper method clicks on a text field or combobox at the point
specified by the x and y coordinates passed in.

Additionally, in several function calls in this file, the y coordinate
has been changed from 63.0 to 62.0 to be more consistent with the other
function calls in the file.

BUG=chromium:59266

Change-Id: I96fb807445b812cbe45a5c1b0631b744727a6f6d