Add a build target and a proper export header for shared library builds.
This CL adds support for Chromium's component build feature to pdfium.

The export header stub in fpdfview.h is expanded to match Chromium's
export mechanisms and patterns (fixes pdfium:825).

A component/shared library build can be triggered by adding

is_component_build = true

as a gn argument. Please note that setting this will also affect some
of pdfiums dependencies like v8, which will be build as components
too.

Additionally, this CL provides a "pdf_source_set" template which
dynamically enables the use of "source_set" when building a complete
static library or a shared library to reduce build time.

When testing this it is recommended to only build the pdfium target as
most of pdfiums test rely on non-public functions which aren't exported
by the shared library.

Bug: pdfium:825,pdfium:826
Change-Id: Icedc538ec535e11d1e53c4d5fabc8c064b275752