Unify of saving documents.
In the original code the method of writing of objects depends on a much unpredictable factors:
as:
1) Is there an updated version of the at least one object in the document.
2) The password is changed.
3) Was this object loaded earlier.
4) The Object is compressed and document have a password.

With these factors it is difficult to predict what will be the final file.
I think to reduce volatility we need to use only one method that works in all cases mentioned.
This method is parse then serialize.

Change-Id: I3d7dcadd10abffbad68d1f993f2dd60b039ed989