Replace raw value for constant error value in string operations
Currently Find() and other methods that return a FX_STRSIZE return -1
to indicate error/failure. This means that there is a lot of magic
numbers and magic checks floating around. The standard library for
similar operations uses a npos constant. This CL implements
FX_STRNPOS, and replaces usages of magic number checking. It also does
some type cleanup along the way where it was obvious that FX_STRSIZE
should be being used.  raw

Removing the magic numbers should make eventually changing FX_STRSIZE
to be unsigned easier in the future.

BUG=pdfium:828

Change-Id: I67e481e44cf2f75a1698afa8fbee4f375a74c490