Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Does this actually fix anything we've seen in the wild?
> 
> Also, do we want to someday move to a model where insert('x', NPOS) means I don't want to care how long the string is, stick 'x' at the end of it ???

Given that I didn't have to change anything other then the unit test call sites, I don't think this fixes any existing bugs. It is mostly removing the possibility of this method covering up bugs in code that calls it in the future.

I am not a big fan of having magic parameters for methods, since we can't make NPOS some sort of non-calculable atom, so there is always the possibility of someone accidentally passing it in. I would prefer to add InsertBefore/InsertAfter (or possibly Before/After) as convenience methods for when you want to do Insert(0, foo) and Insert(end, foo). This makes the intent of the call clear to the reader and it follows the pattern established by Mid/Left/Right. Does that sound like a good idea to you?

Patch-set: 1
Reviewer: Ryan Harrison <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0