Remove unused FromUnicode method
A helper method for FromUnicode existed that converted wchar_t* to
CFX_WideString before calling the remaining FromUnicode. This method
is not used in the code base, so it has been removed. This simplifies
converting FX_STRSIZE to be unsigned, since one of parameters to this
method was of type FX_STRSIZE.

BUG=pdfium:828

Change-Id: I3da677b84ac3b82ba48497e26d2ac80dd14302e6