Update patch set 14
Patch Set 14:

Is the plan to use this for the post processing steps as in the larger CL? Do we know if doing all the decoding as extra walks at the end will cause performance issues or not?

Are there plans other then those post-processing steps in which to use these classes?

Patch-set: 14
Reviewer: dsinclair <5965@fca2d64a-0420-3532-b200-12848c95a47b>