Add checks of index operations on string classes
Specifically the index parameter passed in to GetAt(), SetAt() and
operator[] are now being tested to be in bounds, and adjusted if they
need be. The ASSERTS are included to alert developers to potential
issues with code using these methods, and the internal adjustments are
to prevent out of bounds address accesses in production code.

BUG=chromium:752480, pdfium:828

Change-Id: I9e94d58c98a8eaaaae53cd0e3ffe2123ea17d8c4