LCMS: upgrade to 2.8
This CL upgrades LCMS from version 2.6 to 2.8. The history of each file
was checked to find bits changed that were not in patch files. The
folder was renamed to something more appropriate and that will not need
to be later renamed. The new patch files were applied and numbered
before existing patches. The file cmserr.c was looked at separately
because it is where the memory management methods are defined.

Change-Id: Ic1d84e54803ef9e6b280ef7619bbf0b757312fbf