Update patch set 9
Patch Set 9:

Argh, and I've completely misread the logic.  It needs to go back kinda the way it was, with wchar outside the block.  Take a close look at what happens with an empty string and but an item

Patch-set: 9
Reviewer: Tom Sepez <5495@fca2d64a-0420-3532-b200-12848c95a47b>