Fix use_system_freetype builds for FreeType changes
FreeType has changed so it doesn't define ft_adobe_glyph_list once for
every translation unit, which avoids wasting space in the data segment
with duplicates. However Chromium builds with use_system_freetype true
were depending on this behavior. This fixes these builds while avoiding
duplicate arrays in other builds.

This change makes no difference to pdfium itself, but is necessary when
it is pulled in to Chromium.

This is a follow-on to a previous change that fixed this issue for
component builds.

Change-Id: I3e68d7b778475e4a2781d77cd413e1bc22f82830