Convert string Find methods to return a pair
The Find and ReverseFind methods for WideString, WideStringC,
ByteString, and ByteStringC have been converted from returning a raw
FX_STRSIZE, to returning {bool, FX_STRSIZE}. The boolean indicates
whether or not succeeded.

This allows for removing FX_STRNPOS and by association makes the
conversion of FX_STRSIZE to size_t easier, since it forces checking
the return value of Find to be explictly done as well as taking the
error value out of the range of FX_STRSIZE.

New Contains methods have been added for cases where the success or
failure is all the call site to Find cared about, and the actual
position was ignored.

BUG=pdfium:828

Change-Id: Id827e508c8660affa68cc08a13d96121369364b7