Avoid annotation popups to be clipped at page edges
Currently, annotation popups are placed at a fixed location relative to
the annotation. If an annotation is too much on the right or the bottom
of the page, then its popup will be clipped at the edge of the page.

This CL translates the popups appropriately to avoid getting clipped at
page edges. Also, if an annotation is at the bottom-right corner of the
page, then the popup is placed above and to the left of the annotation,
instead of below and to the right of it.

See a set of comparison screenshots here:
https://drive.google.com/drive/folders/0B7Fvv7JszRyGZXZXR05yVEFWSlE?usp=sharing

Bug=645367,pdfium:750

Change-Id: Ie048281e965523c292dcc704b51188d794f8877b