Do not QuickFloor on cmsintrp
In this CL, some dangerous quickfloors are removed. fclamp() returns a
float between 0.0f and 1.0f, and Domain is unsigned and represents the
number of nodes minus 1. So prefer to just cast to int, which is faster
and will not cause heap-buffer-overflow. In the testcase for the bug,
Input[2] is a number very close but smaller than 1 such that quickfloor
returned 1 (whereas Input[2] >= 1.0 was false).

Bug: chromium:752725
Change-Id: Ibb1763aa120a600e86602f1a46c4cd6d0d6bebd5