Bypass asserts in string [] operator for empty strings
The previous behaviour of [] on an empty string was to return 0
regardless of the index. We wanted to make this more strict, hence the
current behaviour. This has led to a lot of ClusterFuzz issues due to
code depending on the old behaviour. Reverting to the old behaviour
until we have time to correct the call sites using empty strings.

Bug=chromium:752480, pdfium:828

Change-Id: I511eea4148de85bf7f4694351e7a030b1a37f0de