Corrected the error setting after calling FPDF_LoadDocument()
Currently, if FPDF_LoadDocument() is called on a nonexistent file, the
returned FPDF_DOCUMENT will be null, but FPDF_GetLastError() will return
an error of FPDF_ERR_SUCCESS, i.e. no error.

This CL corrects this behavior by updating the error to be FPDF_ERR_FILE
when the file doesn't exist.

Bug=pdfium:452

Change-Id: I3c3ec3a64e049636ddfb2ba5cb5f2745a0e19b6b