Update patch set 4
Patch Set 4:

Mac has had the Y-axis going both up and down for years, so there is a precedent. I'm not sure I can justify why, though.

I do find it confusing when the rectangle is upside down, the font size is negative (inverting its direction), and the matrix scale is negative, and somehow everything cancels out and draws upright. Since this appears to be allowed by PDF itself, there doesn't seem to be much one can do about it.

As long as it is documented, I guess it doesn't matter which way the rectangles go, or whether there are multiple conventions. I can imagine that changing the code to adhere to a new convention could be very tricky with not much to show for it.

Cary

Patch-set: 4
Tag: mailMessageId=<CAE4e=fdwsox3qUEDwyuJ8i-N4aOLf3vyhzvO5dDOLDieGWyUEA@mail.gmail.com>