Convert string Find methods to return a pair
The Find and ReverseFind methods for WideString, WideStringC,
ByteString, and ByteStringC have been converted from returning a raw
FX_STRSIZE, to returning {bool, FX_STRSIZE}. The boolean indicates
whether or not something was found.

This allows for removing FX_STRNPOS and by association makes the
conversion of FX_STRSIZE to size_t easier.  Since removes an error 
value from the range of FX_STRSIZE. It also forces explicit checking
of the result, which makes the call less error prone.

New Contains methods have also been added for cases where the success 
or failure is all the call site cared about, and the actual
position was ignored.

BUG=pdfium:828

Change-Id: Id827e508c8660affa68cc08a13d96121369364b7