Convert FX_STRSIZE to size_t
Change the underlying type for FX_STRSIZE to size_t from int. This
will make the value unsigned and thus all values in the range of the
type will be valid. This allows for the final remove of negative
length strings, but also introduces a bunch of type casting errors,
because many parts of the code base assumed that FX_STRSIZE is int or
another signed type.

BUG=pdfium:828

Change-Id: I231dca59e96fc9330cbb099eecbdfc41fcf86f5b