Calculate AP for multi-line markup annotations correctly
Currently, when constructing AP for multi-line markup annotations, we
only take into account the first set of quadpoints, resulting in only
the first line of the annotation being displayed if the annotation spans
multiple lines.

This CL takes into account all the quadpoints, so multi-line annotations
can be displayed correctly.

Bug=pdfium:876

Change-Id: I23c7e7216c4a69c8cc42e5d6469bfd2ad0ae70db