Update patch set 12
Patch Set 12:

(12 comments)

Everywhere I comment that a change can be in a separate CL, I mean one or a few CLs for all changes, not one for each change. This is just to make flipping the enum a minimal CL as it has a very high change of being reverted.

Patch-set: 12
Reviewer: Henrique Nakashima <7035@fca2d64a-0420-3532-b200-12848c95a47b>