Prepare for converting FX_STRSIZE int->size_t
When turning on this conversion a number of typing issues and other nits where
found in the code base that can be merged in without actually changing the
underlying type. Landing these changes before the type change CL, since there is
a high likelihood that the type change will need to be rolled back, since it is
high risk.

BUG=pdfium:828

Change-Id: I587443d9090055963446485a1aacb8772eb5ca64