Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > I thought we want to do serialization like in chromium JsonWriter
> > > See
> > > https://cs.chromium.org/chromium/src/base/json/json_writer.cc?type=cs
> > 
> > I can see where you're coming from.  As I see it, base::Value is more than just a representation returned from JSON. I believe there are other ways to parse and get these from the system, so separating out the internals from the various print/parse makes sense.
> > 
> > For PDF objects, we're never going to serialize these in a format other than
> > the pdf-style << >>  etc, so keeping them bound to a particular print/parse representation is feasible.
> 
> ok,
> What about "stack overflow" on recursion calls of WriteTo by arrays/dictionaries/streams
> May be i just merge this methods, and remove recursion?

This is a write, and unlikely to occur on any real-world cases since the document already exists. Do you have such a case outside of a unit test? 
On the read side, we limit the parse depth to deal with hostile documents, but I'm not sure this will be an issue here.

Patch-set: 1
Reviewer: Tom Sepez <5495@fca2d64a-0420-3532-b200-12848c95a47b>