Update patch set 2
Patch Set 2:

Can you file a bug for this one? Would like to see why it's needed and how we're going to show it's an improvement?

I'm concerned that the maintenance cost of the extra suffix tree is going to be high. (As well, if we copy all the data into the suffix tree, won't we end up using a lot more memory?)

Patch-set: 2
Reviewer: dsinclair <5965@fca2d64a-0420-3532-b200-12848c95a47b>