Convert selection to use count instead of end index
This CL changes the Text Edit Engine code to use a count instead of an
end index for the selection range. Using count lets us differentiate a
selection at the beginning of 1 character and an empty selection.

A few new tests were added to test unicode word break behaviour, some are not
working yet and are commented out.

Change-Id: Icce8f5003102ef0a850151ccdf16d3c2226d94bf