Move decompressing of JPX out of Init and into Decode
In the existing implementation of the JPX decoder, Init extracts the
header from the image and then immediately decompresses it. This is
problematic if it is a very large image that we won't be able to
allocate a bitmap for. The code has been changed to instead delay
decompression until the Decode method, since things like dest Bitmap
generation can be performed using just the header information.

There is also a bit of renaming/casting cleanup, because I was having
a hard time parsing what was a local vs member variable.

BUG=chromium:761005

Change-Id: I55a55c0be2f88a5352a6ca056c2a816137d7c749