Shift indexing by one in IsHyphen to prevent asserts in [] operator
The existing code did end of range checks by making sure that the value was
never less then 0. This isn't correct when using an unsigned type, since 0 - 1
will wrap arround to the max possible value, and thus still be less then 0. The
pattern that we have been using to resolve this, is shifting the indexing range
by 1, so [1, foo + 1] instead of [0, foo] and applying - 1 where the index is
being used.

BUG=chromium:763256

Change-Id: Ib8bf5ca0e29e73724c4a1c4781362e8a8fc30149