Add guard against reading more then expected from the page
This really shouldn't ever happen, but there used to be this guard in
this code and I am getting reports of crashes after it was removed. I
have added an assert, so hopefully if it is actually occuring, then we
might get a reproduction case based on a debug build crash.

BUG=chromium:763369

Change-Id: Ifaebfbcb0413a1d7777222ba838aaee234f94ae3