Rewrite IsHyphen using string operations
The existing code did end of range checks by making sure that the 
value was never less then 0. This isn't correct when using an unsigned
type, since 0 - 1 will wrap around to the max possible value, and
thus still be less then 0. Additionally the existing code was hard to
follow due to the complexity of some of the low level operations being 
performed.

It has been rewritten using higher level string operations to make it
clearer and correct.

BUG=chromium:763256

Change-Id: Ib8bf5ca0e29e73724c4a1c4781362e8a8fc30149