Revert moving JPX library decode from Init to Decode
Due to some of the size paramters for allocating space in decode
depending on the values produced by opj_decode, this change was
causing misallocation of space, which intern was coausing illegal
reads/writes.

The issue with excessive memory usage that the original CL was trying
to change are less significant then the above mentioned problems, so
reverting this fix and looking for another solution to the
problem. This will re-open bugs https://crbug.com/754423 and
https://crbug.com/761005.

BUG=chromium:764177,chromium:754423

Change-Id: I1cafac8a8117ec1e3bc32b31196bdec719d46477