Fix decoding of JBIG2 integers for values with large magnitude.
The JBIG2 spec says that 32 bits are used to encode (val - 4436)
and does not define limits for the val, so it could be up to
(2^32 + 4435). On the negative side, 32 bits are used to encode
(-val - 4436), making -(2^32 + 4436) the lowest value
representable.

By using a 32-bit integer to store the magnitude, we only supported
up to 2^31 - 1. Larger values became negative. The lowest value we
decoded correctly was -(2^31 - 1).

The typing changes should accommodate these new values, though the
method signature still requires casting down and losing the range
above 2^31 - 1 and below -2^31.

Bug: chromium:761666
Change-Id: I00f5a3abadca51f9bedc5e5d78f7f184040c2f33